refactored smurfs components
