+++ smurfs/src/components/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import './App.css';
 
-import {SmurfsList} from "./";
+import SmurfsList from "./SmurfsList";
 /*
  to wire this component up you're going to need a few things.
  I'll let you do this part on your own.
